requirements.txt
asgiref==3.4.1
Django==3.2.17
django-crispy-forms==1.13.0
django-environ==0.9.0
django-filter==21.1
djangorestframework==3.14.0
opencv-python-headless==4.7.0.68
Pillow==8.4.0
psycopg2-binary==2.9.5
python-barcode==0.14.0
pytz==2022.7.1
pyzbar==0.1.9
reportlab==3.6.8
sqlparse==0.4.3
typing_extensions==4.1.1
